+++ chromium/build/android/pylib/local/emulator/avd.py
@@ -9,7 +9,6 @@ import os
 import socket
 import stat
 import subprocess
-import textwrap
 import threading
 
 from google.protobuf import text_format  # pylint: disable=import-error
@@ -20,11 +19,19 @@ from devil.utils import cmd_helper
 from devil.utils import timeout_retry
 from py_utils import tempfile_ext
 from pylib import constants
+from pylib.local.emulator import ini
 from pylib.local.emulator.proto import avd_pb2
 
 _ALL_PACKAGES = object()
-_DEFAULT_AVDMANAGER_PATH = os.path.join(constants.ANDROID_SDK_ROOT, 'tools',
-                                        'bin', 'avdmanager')
+_DEFAULT_AVDMANAGER_PATH = os.path.join(
+    constants.ANDROID_SDK_ROOT, 'cmdline-tools', 'latest', 'bin', 'avdmanager')
+# Default to a 480dp mdpi screen (a relatively large phone).
+# See https://developer.android.com/training/multiscreen/screensizes
+# and https://developer.android.com/training/multiscreen/screendensities
+# for more information.
+_DEFAULT_SCREEN_DENSITY = 160
+_DEFAULT_SCREEN_HEIGHT = 960
+_DEFAULT_SCREEN_WIDTH = 480
 
 
 class AvdException(Exception):
@@ -72,12 +79,13 @@ class _AvdManagerAgent(object):
 
     self._env = dict(os.environ)
 
-    # avdmanager, like many tools that have evolved from `android`
-    # (http://bit.ly/2m9JiTx), uses toolsdir to find the SDK root.
+    # The avdmanager from cmdline-tools would look two levels
+    # up from toolsdir to find the SDK root.
     # Pass avdmanager a fake directory under the directory in which
     # we install the system images s.t. avdmanager can find the
     # system images.
-    fake_tools_dir = os.path.join(self._sdk_root, 'non-existent-tools')
+    fake_tools_dir = os.path.join(self._sdk_root, 'non-existent-tools',
+                                  'non-existent-version')
     self._env.update({
         'ANDROID_AVD_HOME':
         self._avd_home,
@@ -196,7 +204,7 @@ class AvdConfig(object):
       cipd_json_output: string path to pass to `cipd create` via -json-output.
     """
     logging.info('Installing required packages.')
-    self.Install(packages=[
+    self._InstallCipdPackages(packages=[
         self._config.emulator_package,
         self._config.system_image_package,
     ])
@@ -227,19 +235,30 @@ class AvdConfig(object):
       with open(root_ini, 'a') as root_ini_file:
         root_ini_file.write('path.rel=avd/%s.avd\n' % self._config.avd_name)
 
-      with open(config_ini, 'a') as config_ini_file:
-        config_ini_file.write(
-            textwrap.dedent("""\
-                disk.dataPartition.size=4G
-                hw.lcd.density=160
-                hw.lcd.height=960
-                hw.lcd.width=480
-                """))
+      if os.path.exists(config_ini):
+        with open(config_ini) as config_ini_file:
+          config_ini_contents = ini.load(config_ini_file)
+      else:
+        config_ini_contents = {}
+      height = (self._config.avd_settings.screen.height
+                or _DEFAULT_SCREEN_HEIGHT)
+      width = (self._config.avd_settings.screen.width or _DEFAULT_SCREEN_WIDTH)
+      density = (self._config.avd_settings.screen.density
+                 or _DEFAULT_SCREEN_DENSITY)
+
+      config_ini_contents.update({
+          'disk.dataPartition.size': '4G',
+          'hw.lcd.density': density,
+          'hw.lcd.height': height,
+          'hw.lcd.width': width,
+      })
+      with open(config_ini, 'w') as config_ini_file:
+        ini.dump(config_ini_contents, config_ini_file)
 
       # Start & stop the AVD.
       self._Initialize()
-      instance = _AvdInstance(self._emulator_path, self._config.avd_name,
-                              self._emulator_home)
+      instance = _AvdInstance(self._emulator_path, self._emulator_home,
+                              self._config)
       instance.Start(read_only=False, snapshot_save=snapshot)
       device_utils.DeviceUtils(instance.serial).WaitUntilFullyBooted(
           timeout=180, retries=0)
@@ -304,11 +323,19 @@ class AvdConfig(object):
         avd_manager.Delete(avd_name=self._config.avd_name)
 
   def Install(self, packages=_ALL_PACKAGES):
-    """Installs the requested CIPD packages.
+    """Installs the requested CIPD packages and prepares them for use.
+
+    This includes making files writeable and revising some of the
+    emulator's internal config files.
 
     Returns: None
     Raises: AvdException on failure to install.
     """
+    self._InstallCipdPackages(packages=packages)
+    self._MakeWriteable()
+    self._EditConfigs()
+
+  def _InstallCipdPackages(self, packages):
     pkgs_by_dir = {}
     if packages is _ALL_PACKAGES:
       packages = [
@@ -328,9 +355,9 @@ class AvdConfig(object):
         os.makedirs(cipd_root)
       ensure_path = os.path.join(cipd_root, '.ensure')
       with open(ensure_path, 'w') as ensure_file:
-        # Make CIPD ensure that all files are present, even if
-        # it thinks the package is installed.
-        ensure_file.write('$ParanoidMode CheckPresence\n\n')
+        # Make CIPD ensure that all files are present and correct,
+        # even if it thinks the package is installed.
+        ensure_file.write('$ParanoidMode CheckIntegrity\n\n')
         for pkg in pkgs:
           ensure_file.write('%s %s\n' % (pkg.package_name, pkg.version))
           logging.info('  %s %s', pkg.package_name, pkg.version)
@@ -351,6 +378,7 @@ class AvdConfig(object):
                                                        str(e)),
             command=ensure_cmd)
 
+  def _MakeWriteable(self):
     # The emulator requires that some files are writable.
     for dirname, _, filenames in os.walk(self._emulator_home):
       for f in filenames:
@@ -359,6 +387,35 @@ class AvdConfig(object):
         if mode & stat.S_IRUSR:
           mode = mode | stat.S_IWUSR
         os.chmod(path, mode)
+
+  def _EditConfigs(self):
+    android_avd_home = os.path.join(self._emulator_home, 'avd')
+    avd_dir = os.path.join(android_avd_home, '%s.avd' % self._config.avd_name)
+
+    hardware_qemu_path = os.path.join(avd_dir, 'hardware-qemu.ini')
+    if os.path.exists(hardware_qemu_path):
+      with open(hardware_qemu_path) as hardware_qemu_file:
+        hardware_qemu_contents = ini.load(hardware_qemu_file)
+    else:
+      hardware_qemu_contents = {}
+
+    hardware_qemu_contents['hw.sdCard'] = 'true'
+    if self._config.avd_settings.sdcard.size:
+      sdcard_path = os.path.join(avd_dir, 'cr-sdcard.img')
+      if not os.path.exists(sdcard_path):
+        mksdcard_path = os.path.join(
+            os.path.dirname(self._emulator_path), 'mksdcard')
+        mksdcard_cmd = [
+            mksdcard_path,
+            self._config.avd_settings.sdcard.size,
+            sdcard_path,
+        ]
+        cmd_helper.RunCmd(mksdcard_cmd)
+
+      hardware_qemu_contents['hw.sdCard.path'] = sdcard_path
+
+    with open(hardware_qemu_path, 'w') as hardware_qemu_file:
+      ini.dump(hardware_qemu_contents, hardware_qemu_file)
 
   def _Initialize(self):
     if self._initialized:
@@ -389,8 +446,7 @@ class AvdConfig(object):
       An _AvdInstance.
     """
     self._Initialize()
-    return _AvdInstance(self._emulator_path, self._config.avd_name,
-                        self._emulator_home)
+    return _AvdInstance(self._emulator_path, self._emulator_home, self._config)
 
   def StartInstance(self):
     """Starts an AVD instance.
@@ -410,15 +466,16 @@ class _AvdInstance(object):
   but its other methods can be freely called.
   """
 
-  def __init__(self, emulator_path, avd_name, emulator_home):
+  def __init__(self, emulator_path, emulator_home, avd_config):
     """Create an _AvdInstance object.
 
     Args:
       emulator_path: path to the emulator binary.
-      avd_name: name of the AVD to run.
       emulator_home: path to the emulator home directory.
+      avd_config: AVD config proto.
     """
-    self._avd_name = avd_name
+    self._avd_config = avd_config
+    self._avd_name = avd_config.avd_name
     self._emulator_home = emulator_home
     self._emulator_path = emulator_path
     self._emulator_proc = None
@@ -428,8 +485,13 @@ class _AvdInstance(object):
   def __str__(self):
     return '%s|%s' % (self._avd_name, (self._emulator_serial or id(self)))
 
-  def Start(self, read_only=True, snapshot_save=False, window=False):
+  def Start(self,
+            read_only=True,
+            snapshot_save=False,
+            window=False,
+            writable_system=False):
     """Starts the emulator running an instance of the given AVD."""
+
     with tempfile_ext.TemporaryFileName() as socket_path, (contextlib.closing(
         socket.socket(socket.AF_UNIX))) as sock:
       sock.bind(socket_path)
@@ -439,11 +501,16 @@ class _AvdInstance(object):
           self._avd_name,
           '-report-console',
           'unix:%s' % socket_path,
+          '-no-boot-anim',
       ]
+
       if read_only:
         emulator_cmd.append('-read-only')
       if not snapshot_save:
         emulator_cmd.append('-no-snapshot-save')
+      if writable_system:
+        emulator_cmd.append('-writable-system')
+
       emulator_env = {}
       if self._emulator_home:
         emulator_env['ANDROID_EMULATOR_HOME'] = self._emulator_home
@@ -454,6 +521,7 @@ class _AvdInstance(object):
           raise AvdException('Emulator failed to start: DISPLAY not defined')
       else:
         emulator_cmd.append('-no-window')
+
       sock.listen(1)
 
       logging.info('Starting emulator.')
@@ -484,9 +552,13 @@ class _AvdInstance(object):
     """Stops the emulator process."""
     if self._emulator_proc:
       if self._emulator_proc.poll() is None:
-        self._emulator_proc.terminate()
+        if self._emulator_serial:
+          device_utils.DeviceUtils(self._emulator_serial).adb.Emu('kill')
+        else:
+          self._emulator_proc.terminate()
         self._emulator_proc.wait()
       self._emulator_proc = None
+
     if self._sink:
       self._sink.close()
       self._sink = None
